apps/requirements.txt
absl-py==2.1.0
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.0.4
Babel==2.14.0
beautifulsoup4==4.12.3
billiard==4.2.0
bleach==6.1.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cli-helpers==2.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
configobj==5.0.8
contourpy==1.2.0
coverage==7.4.1
cryptography==42.0.1
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2
django-allauth==0.60.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==23.5
django-jazzmin==2.6.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
execnet==2.0.2
executing==2.0.1
factory-boy==3.3.0
Faker==22.6.0
fastjsonschema==2.19.1
flake8==7.0.0
flower==2.0.1
fonttools==4.50.0
fqdn==1.5.1
h11==0.14.0
h5py==3.10.0
htmlmin==0.1.12
httpcore==1.0.4
httpx==0.27.0
humanize==4.9.0
idna==3.6
ImageHash==4.3.1
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
keras==3.1.1
kiwisolver==1.4.5
kombu==5.3.5
lightgbm==4.3.0
llvmlite==0.42.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
model-bakery==1.17.0
multimethod==1.11.2
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.1.2
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
optree==0.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandas-profiling==3.6.6
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
pendulum==2.1.2
pexpect==4.9.0
pgcli==4.0.1
pgspecial==2.1.1
phik==0.12.4
pillow==10.2.0
platformdirs==4.2.0
plotly==5.20.0
pluggy==1.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg==3.1.17
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.0.0
pytest-django==4.8.0
pytest-factoryboy==2.6.0
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.4
pytzdata==2020.1
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.17.1
scikit-learn==1.4.1.post1
scipy==1.11.4
seaborn==0.12.2
Send2Trash==1.8.2
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
statsmodels==0.14.1
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.4.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typeguard==4.1.5
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
visions==0.7.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wordcloud==1.9.3
xgboost==2.0.3
ydata-profiling==4.7.0
